// client/src/redux/posts/postTypes.ts
export enum postTypes {
    CREATE_POST_REQUEST = "CREATE_POST_REQUEST",
    CREATE_POST_SUCCESS = "CREATE_POST_SUCCESS",
    CREATE_POST_FAILURE = "CREATE_POST_FAILURE",
    CHANGE_CURRENT_POST = "CHANGE_CURRENT_POST",
    RESET_CURRENT_POST = "RESET_CURRENT_POST",
    FETCH_POSTS_REQUEST = "FETCH_POSTS_REQUEST",
    FETCH_POSTS_SUCCESS = "FETCH_POSTS_SUCCESS",
    FETCH_POSTS_FAILURE = "FETCH_POSTS_FAILURE",
}